Testing with Yalc
I removed the `pnpm-workspace.yaml`, then followed the documented Yalc
workflow, first to build and "publish" `fiber`, then `drei`, then to build the playground.

I've modified the playground to try to demonstrate the `DirectionalLight` issue...
but the lights seem to be working.  I use the "For" example to get a good, stable look.
Note that I changed the box to a dodecahedron to make the lighting more obvious.