Better lighting defect demo
I have added a PerspectiveCamera from drei, and put the lighting inside that.  This is
much more faithful to what my own app is doing.  Use the "For" example, and notice the yellow shade of the shapes, but that it becomes more red vs. green as you zoom in.
The lighting should always look red vs. green if the transformation of the lighting is
correct inside the PerspectiveCamera.

In my app, all three lights end up acting like headlamps because the distance is greater,
but it is the same effect.