initial commit minimal branch
Implementation of proxy-based threejs renderer, based on the exploration done in the zustand-to-store branch: see https://github.com/solidjs-community/solid-three/pull/12

This branch deviates from the previous goal of staying close to r3f's codebase in favor of simplification and 'solidification'.

How the props are managed (= applying jsx-props to three-instand and managing of the scene graph) and the Primitive-component are copied from the zustand-to-store branch.

differences:
- Custom event handling
- Augment the threejs instance with far less in comparison then r3f: we only store the props of the component